middleware error handling

+++ src/service/user.service.ts
@@ -3,6 +3,7 @@ import {UserModel} from "../models/user.model";
 import {PrismaClient, token} from '@prisma/client';
 import {tokenService} from "./token.service";
 import UserDto from '../dtos/user.dto';
+import {ApiError} from "../errors/api.error";
 
 const prisma = new PrismaClient();
 
@@ -16,7 +17,7 @@ export const userService = {
         });
 
         if (candidate) {
-            throw new Error("User already exists")
+            throw ApiError.BadRequest("User already exists");
         }
 
         const hashedPassword = hashSync(password, 10);
@@ -39,28 +40,32 @@ export const userService = {
         });
 
         if (user == null) {
-            throw new Error('UserModel not found');
+            throw ApiError.BadRequest('User not found');
         }
 
         if (!compareSync(password, user.password)) {
-            throw new Error('Incorrect password');
+            throw ApiError.BadRequest('Incorrect password');
         }
 
         return await saveToken(user);
     },
 
     async logout(refreshToken: string) : Promise<token> {
+        if (!refreshToken) {
+            throw ApiError.UnauthorizedError();
+        }
         return await tokenService.removeToken(refreshToken);
     },
 
     async refresh(refreshToken: string) : Promise<{refreshToken: string, accessToken: string, userDTO: UserDto}> {
         if (!refreshToken) {
-            throw new Error("Unauthorized error");
+            throw ApiError.UnauthorizedError();
         }
         const userId = tokenService.validateRefreshToken(refreshToken);
         const token = await tokenService.findToken(refreshToken);
+
         if (!userId || !token) {
-            throw new Error("Unauthorized error");
+            throw ApiError.UnauthorizedError();
         }
 
         const user = await prisma.user.findUnique({
@@ -70,7 +75,7 @@ export const userService = {
         });
 
         if (!user) {
-            throw new Error("User not found");
+            throw ApiError.BadRequest("User not found");
         }
 
         return await saveToken(user);
